# Conflicts:
#	NaoTHSoccer/Source/Cognition/Cognition.cpp

+++ Utils/CollectStick/startBrainwashing.sh
@@ -29,6 +29,7 @@ exec_cmd_and_return_or_default() {
 
 
 # set file vars
+currentDir=$(pwd)  # /media/brainwasher
 infoFile="/home/nao/Config/nao.info"
 errorFile="/home/nao/brainwasher.log"
 
@@ -36,7 +37,7 @@ errorFile="/home/nao/brainwasher.log"
 current_date=$(date +"%y%m%d-%H%M")
 current_nao=$(sed -n "2p" $infoFile)
 current_nao_name=$(cat /etc/hostname) # get the name, eg. "nao96"
-current_nao_number=$(cat /etc/hostname | grep -Eo "[0-9]{2}") # get the number, e.g. "96"
+current_nao_number=$(cat /etc/hostname | grep -Eo "[0-9]{2}" | head -n 1 ) # get the number, e.g. "96"
 current_nao_player=$(exec_cmd_and_return_or_default 'grep -o -E "\[PlayerInfo\] playerNumber.*" /var/log/messages | tail -1 | sed "s/[^0-9]*//g"' '0') # grep the playerNumber from log, take the last one and extract the actual number
 
 # if no player number could be extracted, set it to default
@@ -72,7 +73,7 @@ logger "Brainwasher:copy files"
 # create directory
 #dir_name=$current_date-$current_nao
 dir_name=${current_nao_player}_${current_nao_number}_${current_nao}_${current_date}
-target_path=/media/brainwasher/$dir_name
+target_path=$currentDir/$dir_name
 mkdir -p $target_path
 
 # copy info file of the nao
@@ -86,33 +87,38 @@ echo "$current_compile_time" >> $target_path/nao.info
 echo "$current_compile_owner" >> $target_path/nao.info
 
 # find log files and copy them to the created directory
-#find -L /tmp -type d -name media -prune -o -name "*.log" -exec cp {} /media/brainwasher/$current_date-$current_nao \;
+#find -L /tmp -type d -name media -prune -o -name "*.log" -exec cp {} $currentDir/$current_date-$current_nao \;
 # find log files, create MD5 hashes and copy everything to the created directory
 for f in $(find -L /tmp -type d -name media -prune -o -name "*.log")
 do
-	md5sum $f | sed -e "s/\/tmp\///g" > "$f.md5"
+	md5sum $f | sed -e "s#/tmp/##g" > "$f.md5"
 	cp "$f.md5" $f $target_path/
 	check_for_errors "Brainwasher:ERROR copying $f"
 done
 
 for f in $(find -L /dev/shm -type d -name media -prune -o -name "*.log")
 do
-	md5sum $f | sed -e "s/\/dev/shm\///g" > "$f.md5"
+	md5sum $f | sed -e "s#/dev/shm/##g" > "$f.md5"
 	cp "$f.md5" $f $target_path/
 	check_for_errors "Brainwasher:ERROR copying $f"
 done
 
 # copy the config directory
+logger "copy the config directory"
 cd /home/nao/naoqi
 zip -q -r -0 $target_path/config.zip Config
 cd -
 check_for_errors "Brainwasher:ERROR copying config"
 
 # copy logs of naoth binary (std::out/::err) and clear them afterwards
-cp "/var/log/naoth.log" "/var/log/naoth_err.log" $target_path/
-check_for_errors "Brainwasher:ERROR copying /var/log/naoth.log"
-> /var/log/naoth.log
-> /var/log/naoth_err.log
+if [ -f "/var/log/naoth.log" ]; then
+	cp "/var/log/naoth.log" "/var/log/naoth_err.log" $target_path/
+	check_for_errors "Brainwasher:ERROR copying /var/log/naoth.log"
+	> /var/log/naoth.log
+	> /var/log/naoth_err.log
+else
+	logger "naoth.log does not exist!"
+fi
 
 # find and copy trace dump files since boot and log errors
 find /home/nao -maxdepth 1 -type f -mmin -$current_boot_time -iname "trace.dump.*" -exec zip -q -0 $target_path/dumps.zip {} + 2> $errorFile
